requirements.txt
aiogram==3.17.0
python-dotenv
python-telegram-bot==20.7
python-dotenv==1.0.0
SpeechRecognition==3.10.0
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
pydub==0.25.1
